requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.5
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.14.0
et-xmlfile==1.0.1
gunicorn==20.1.0
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.4.0
pip==22.3.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.7
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
wheel==0.38.4
whitenoise==6.3.0
